Clean API layer.
