+++ src/main/java/com/elephasvacation/tms/web/api/RoomCategoryServlet.java
@@ -1,0 +1,56 @@
+/*
+@author : Dhanusha Perera
+@date : 16/07/2021
+*/
+
+package com.elephasvacation.tms.web.api;
+
+
+import com.elephasvacation.tms.web.exception.ResponseExceptionUtil;
+import jakarta.json.bind.Jsonb;
+import jakarta.json.bind.JsonbBuilder;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+@WebServlet(name = "RoomCategoryServlet", urlPatterns = "/api/v1/room-categories/*")
+public class RoomCategoryServlet extends HttpServlet {
+
+    private static final Logger logger = LoggerFactory.getLogger(RoomCategoryServlet.class);
+
+    Jsonb jsonb = JsonbBuilder.create();
+
+    @Override
+    protected void service(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        try {
+            super.service(request, response);
+        } catch (Throwable t) {
+            ResponseExceptionUtil.handle(t, response);
+        }
+    }
+
+    @Override
+    protected void doGet(HttpServletRequest request, HttpServletResponse response) {
+
+    }
+
+    @Override
+    protected void doPost(HttpServletRequest request, HttpServletResponse response) {
+
+    }
+
+    @Override
+    protected void doPut(HttpServletRequest request, HttpServletResponse response) {
+
+    }
+
+    @Override
+    protected void doDelete(HttpServletRequest request, HttpServletResponse response) {
+    }
+}
