--- src/main/java/com/elephasvacation/tms/web/api/accommodation/AccommodationAPIImpl.java
@@ -1,73 +1,0 @@
-/*
- * MIT License
- *
- * Copyright (c) 2021 Dhanusha Perera
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-/*
- * @author : Dhanusha Perera
- * @date : 10/07/2021
- */
-package com.elephasvacation.tms.web.api.accommodation;
-
-import com.elephasvacation.tms.web.business.BOFactory;
-import com.elephasvacation.tms.web.business.BOTypes;
-import com.elephasvacation.tms.web.business.custom.AccommodationBO;
-import com.elephasvacation.tms.web.dto.AccommodationDTO;
-
-import java.sql.Connection;
-import java.util.List;
-
-public class AccommodationAPIImpl implements AccommodationAPI {
-
-    private static AccommodationBO accommodationBO = BOFactory
-            .getInstance()
-            .getBO(BOTypes.ACCOMMODATION);
-
-    @Override
-    public void setConnection(Connection connection) throws Exception {
-        accommodationBO.setConnection(connection);
-    }
-
-    @Override
-    public Integer createAccommodation(AccommodationDTO accommodationDTO) throws Exception {
-        return accommodationBO.createAccommodation(accommodationDTO);
-    }
-
-    @Override
-    public boolean updateAccommodation(AccommodationDTO accommodationDTO) throws Exception {
-        return accommodationBO.updateAccommodation(accommodationDTO);
-    }
-
-    @Override
-    public boolean deleteAccommodation(int accommodationID) throws Exception {
-        return accommodationBO.deleteAccommodation(accommodationID);
-    }
-
-    @Override
-    public AccommodationDTO getAccommodationByID(int accommodationID) throws Exception {
-        return accommodationBO.getAccommodationByID(accommodationID);
-    }
-
-    @Override
-    public List<AccommodationDTO> getAllAccommodations() throws Exception {
-        return accommodationBO.getAllAccommodations();
-    }
-}
