--- src/main/java/com/elephasvacation/tms/web/api/customer/customerAPI/CustomerAPIImpl.java
@@ -1,99 +1,0 @@
-/**
- * MIT License
- * <p>
- * Copyright (c) 2021 Dhanusha Perera
- * <p>
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- * <p>
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- * <p>
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- *
- * @author : Dhanusha Perera
- * @date : 03/07/2021
- * @author : Dhanusha Perera
- * @date : 03/07/2021
- * @author : Dhanusha Perera
- * @date : 03/07/2021
- * @author : Dhanusha Perera
- * @date : 03/07/2021
- * @author : Dhanusha Perera
- * @date : 03/07/2021
- * @author : Dhanusha Perera
- * @date : 03/07/2021
- * @author : Dhanusha Perera
- * @date : 03/07/2021
- * @author : Dhanusha Perera
- * @date : 03/07/2021
- */
-/**
- * @author : Dhanusha Perera
- * @date : 03/07/2021
- */
-package com.elephasvacation.tms.web.api.customer.customerAPI;
-
-import com.elephasvacation.tms.web.business.BOFactory;
-import com.elephasvacation.tms.web.business.BOTypes;
-import com.elephasvacation.tms.web.business.custom.CustomerBO;
-import com.elephasvacation.tms.web.dto.CustomerDTO;
-
-import java.sql.Connection;
-import java.util.List;
-
-public class CustomerAPIImpl implements CustomerAPI {
-
-    private final CustomerBO customerBO;
-    private Connection connection;
-
-
-    public CustomerAPIImpl() {
-        this.customerBO = BOFactory.getInstance().getBO(BOTypes.CUSTOMER);
-    }
-
-    @Override
-    public void setConnection(Connection connection) throws Exception {
-        this.customerBO.setConnection(connection);
-    }
-
-    @Override
-    public List<CustomerDTO> getAllCustomers() throws Exception {
-//        System.out.println("Get All Customers from Database!");
-        return this.customerBO.getAllCustomers();
-    }
-
-    @Override
-    public CustomerDTO getCustomerByID(int customerID) throws Exception {
-//        System.out.println("Get Customer By ID: " + customerID + ", from Database!");
-        return this.customerBO.getCustomerByID(customerID);
-    }
-
-    @Override
-    public Integer createCustomer(CustomerDTO customerDTO) throws Exception {
-//        System.out.println("Create new Customer!");
-        return this.customerBO.createCustomer(customerDTO);
-    }
-
-    @Override
-    public boolean updateCustomer(CustomerDTO customerDTO) throws Exception {
-//        System.out.println("Update new Customer!");
-        return this.customerBO.updateCustomer(customerDTO);
-    }
-
-    @Override
-    public boolean deleteCustomer(int customerID) throws Exception {
-//        System.out.println("Delete Customer by ID!");
-        return this.customerBO.deleteCustomer(customerID);
-    }
-}
